OSS-83: rename http -> api
