+++ core/internal/core/segment/segment_api.go
