+++ core/internal/core/segmentrule/segmentrule_api.go
