+++ core/internal/core/trait/trait_api.go
