+++ core/internal/pkg/api/api_metrics.go
@@ -1,4 +1,4 @@
-package http
+package api
 
 import (
 	"github.com/gin-gonic/gin"
